Checks `<instanceName/>`, edits in place.
Reformatted naming_schemes to be easier to read. Added T/F dict
to explain when qxmledit needs to add <instanceName/> information.

qxmledit checks a few new things: (1) HQ and FQ have the same
locations (2) checks if <instanceName/> is already in XML and if
that makes sense.

qxml improved help description